Atualizando a main com a Mudancas-Marina

+++ VLDocesWeb/Models/ProductModel.cs
@@ -5,7 +5,7 @@ public class Product
     public int Id_Produto { set; get; }
     public string Nome { set; get; }
     public string Descricao { set; get; }
-    public float Preco { set; get; }
+    public decimal Preco { set; get; }
     public int Id_Categoria { set; get; }
     
     //public string URL_Img { set; get; }
